feat: Added BadRequest Responses.

+++ WebApi/Controllers/AdminController.cs
@@ -39,7 +39,8 @@ namespace WebApi.Controllers
         /// <response code="200">Devuelve una lista de usuarios.</response>        
         /// <response code="401">Unauthorized. No se ha indicado o es incorrecto el Token JWT de acceso.</response>               
         /// <response code="403">Forbidden. Token JWT de acceso es correcto pero no tiene autorización para el recurso.</response>               
-        /// <response code="404">NotFound. No se ha encontrado el usuario actual.</response> 
+        /// <response code="500">InternalServerError.Ha habido un error al procesar la peticion</response>
+        /// <returns>Una lista de todos los usuarios</returns>
         [HttpGet]
         [Route("api/Admin/Users")]
         [ResponseType(typeof(List<ApplicationUser>))]
@@ -53,21 +54,25 @@ namespace WebApi.Controllers
             {
                 return InternalServerError(e);
             }
-        } 
-        
+        }
+
         /// <summary>
         /// Obtiene informacion del Usuario con UserName.
         /// </summary>
         /// <param name="userName">Nombre Unico del Usuario a devolver.</param>
-        /// <response code="200">Devuelve el usuario.</response>        
+        /// <response code="200">Devuelve el usuario.</response>
+        /// <response code="400">BadRequest. La petición es incorrecta</response>
         /// <response code="401">Unauthorized. No se ha indicado o es incorrecto el Token JWT de acceso.</response>               
         /// <response code="403">Forbidden. Token JWT de acceso es correcto pero no tiene autorización para el recurso.</response>               
-        /// <response code="404">NotFound. No se ha encontrado el usuario actual.</response> 
+        /// <response code="404">NotFound. No se ha encontrado el usuario actual.</response>
+        /// <response code="500">InternalServerError. Ha habido un error al procesar la peticion</response>
         [HttpGet]
         [Route("api/Admin/Users/{userName}")]
         [ResponseType(typeof(ApplicationUser))]
         public IHttpActionResult Get(string userName)
         {
+            if (userName == null) return BadRequest("UserName cannot be null");
+
             try
             {
                 return Ok(_userRepository.GetUser(userName));
@@ -86,16 +91,19 @@ namespace WebApi.Controllers
         /// Create a new User.
         /// </summary>
         /// <param name="user"></param>
-        /// <response code="200"></response>        
+        /// <response code="201"></response>
+        /// <response code="400">BadRequest. La petición es incorrecta</response>
         /// <response code="401">Unauthorized. No se ha indicado o es incorrecto el Token JWT de acceso.</response>               
-        /// <response code="403">Forbidden. Token JWT de acceso es correcto pero no tiene autorización para el recurso.</response>               
-        /// <response code="404">NotFound. No se ha encontrado el usuario actual.</response>
+        /// <response code="403">Forbidden. Token JWT de acceso es correcto pero no tiene autorización para el recurso.</response>
+        /// <response code="500">InternalServerError.Ha habido un error al procesar la peticion</response>
         /// <returns></returns>
         [HttpPost]
         [Route("api/Admin/Users")]
         [ResponseType(typeof(string))]
         public async Task<IHttpActionResult> Post([FromBody] RegisterModel user)
         {
+            if (user == null) return BadRequest("User cannot be null");
+
             try
             { 
                 return Created("Created new User", await _userRepository.AddUserAsync(user.UserName, user.Email, user.Password, user.InitialBalance));
@@ -104,25 +112,33 @@ namespace WebApi.Controllers
             {
                 return InternalServerError(e);
             }
-        } 
-        
+        }
+
         /// <summary>
         /// Edita un usuario.
         /// </summary>
         /// <param name="user"></param>
-        /// <response code="200"></response>        
+        /// <response code="200"></response>
+        /// <response code="400">BadRequest. La petición es incorrecta</response>
         /// <response code="401">Unauthorized. No se ha indicado o es incorrecto el Token JWT de acceso.</response>               
-        /// <response code="403">Forbidden. Token JWT de acceso es correcto pero no tiene autorización para el recurso.</response>               
+        /// <response code="403">Forbidden. Token JWT de acceso es correcto pero no tiene autorización para el recurso.</response>
         /// <response code="404">NotFound. No se ha encontrado el usuario actual.</response>
+        /// <response code="500">InternalServerError.Ha habido un error al procesar la peticion</response>
         /// <returns>Devuelve el Usuario Editado.</returns>
         [HttpPut]
         [Route("api/Admin/Users")]
         [ResponseType(typeof(ApplicationUser))]
         public IHttpActionResult Put([FromBody] ApplicationUser user)
         {
+            if (user == null) return BadRequest("User Cannot be null");
+
             try
             {   
                 return Ok(_userRepository.EditUser(user));
+            }
+            catch (KeyNotFoundException)
+            {
+                return NotFound();
             }
             catch (Exception e)
             {
@@ -136,68 +152,94 @@ namespace WebApi.Controllers
         /// <param name="amount">Cantidad a añadir</param>
         /// <param name="userName">Destinatario</param>
         /// <response code="200"></response>        
+        /// <response code="400">BadRequest. La petición es incorrecta</response>            
         /// <response code="401">Unauthorized. No se ha indicado o es incorrecto el Token JWT de acceso.</response>               
-        /// <response code="403">Forbidden. Token JWT de acceso es correcto pero no tiene autorización para el recurso.</response>               
+        /// <response code="403">Forbidden. Token JWT de acceso es correcto pero no tiene autorización para el recurso.</response>
         /// <response code="404">NotFound. No se ha encontrado el usuario actual.</response>
+        /// <response code="500">InternalServerError.Ha habido un error al procesar la peticion</response>
         /// <returns>Devuelve el nuevo Balance.</returns>
         [HttpPut]
         [Route("api/Admin/Balance/{userName}/Add")]
         [ResponseType(typeof(string))]
         public IHttpActionResult AddBalance([FromBody] int amount, string userName)
         {
+            if (userName == null) return BadRequest("User cannot be null");
+            if (amount < 0) return BadRequest("Amount cannot be negative");
+
             try
-            {   
+            {
                 return Ok(_userRepository.AddBalance(userName, amount));
+            }
+            catch (KeyNotFoundException)
+            {
+                return NotFound();
             }
             catch (Exception e)
             {
                 return InternalServerError(e);
             }
         }
-        
+
         /// <summary>
         /// Descuenta una cantidad al Balance del Usuario.
         /// </summary>
         /// <param name="amount">Cantidad a descontar</param>
         /// <param name="userName">Destinatario</param>
-        /// <response code="200"></response>        
+        /// <response code="200"></response>
+        /// <response code="400">BadRequest. La petición es incorrecta</response>  
         /// <response code="401">Unauthorized. No se ha indicado o es incorrecto el Token JWT de acceso.</response>               
-        /// <response code="403">Forbidden. Token JWT de acceso es correcto pero no tiene autorización para el recurso.</response>               
+        /// <response code="403">Forbidden. Token JWT de acceso es correcto pero no tiene autorización para el recurso.</response>
         /// <response code="404">NotFound. No se ha encontrado el usuario actual.</response>
+        /// <response code="500">InternalServerError. Ha habido un error al procesar la peticion</response>
         /// <returns>Devuelve el nuevo Balance.</returns>
         [HttpPut]
         [Route("api/Admin/Balance/{userName}/Remove")]
         [ResponseType(typeof(string))]
         public IHttpActionResult RemoveBalance([FromBody] int amount, string userName)
         {
+            if (userName == null) return BadRequest("User cannot be null");
+            if (amount < 0) return BadRequest("Amount cannot be negative");
+
             try
             {   
                 return Ok(_userRepository.RemoveBalance(userName, amount));
+            }
+            catch (KeyNotFoundException)
+            {
+                return NotFound();
             }
             catch (Exception e)
             {
                 return InternalServerError(e);
             }
         }
-        
+
         /// <summary>
         /// Borra un usuario.
         /// </summary>
         /// <param name="userName">Nombre Unico del Usuario a borrar.</param>
-        /// <response code="200"></response>        
+        /// <response code="200"></response>
+        /// <response code="400">BadRequest. La petición es incorrecta</response>  
         /// <response code="401">Unauthorized. No se ha indicado o es incorrecto el Token JWT de acceso.</response>               
         /// <response code="403">Forbidden. Token JWT de acceso es correcto pero no tiene autorización para el recurso.</response>               
         /// <response code="404">NotFound. No se ha encontrado el usuario actual.</response>
+        /// <response code="500">InternalServerError.Ha habido un error al procesar la peticion</response>
         /// <returns></returns>
         [HttpDelete]
         [Route("api/Admin/Users/{userName}")]
         [ResponseType(typeof(string))]
         public IHttpActionResult Delete(string userName)
         {
+            if (userName == null) return BadRequest("User cannot be null");
+
             try
             {
                 _userRepository.DeleteUser(userName);
                 return Ok("User delete Succes");
+            }
+            catch (KeyNotFoundException)
+            {
+                return NotFound();
             }
             catch (Exception e)
             {
